Fix integer pwm issue
I had made the pwm an integer, so there was no way to gradually change it, i.e., by fine movement of a gamepad axis. This is fixed now by reverting the pwm value back to a decimal value and only rounding it to an integer when displaying, storing and sending it.